+++ Common/src/main/java/com/nyfaria/nyfspetshop/client/CommonClientClass.java
@@ -1,0 +1,5 @@
+package com.nyfaria.nyfspetshop.client;
+
+public class CommonClientClass {
+
+}