+++ Fabric/src/main/java/com/nyfaria/nyfspetshop/NyfsPetShopClient.java
@@ -1,0 +1,11 @@
+package com.nyfaria.nyfspetshop;
+
+import net.fabricmc.api.ClientModInitializer;
+
+public class NyfsPetShopClient implements ClientModInitializer {
+    
+    @Override
+    public void onInitializeClient() {
+
+    }
+}