+++ Forge/src/main/java/com/nyfaria/nyfspetshop/event/CommonModEvents.java
@@ -1,0 +1,14 @@
+package com.nyfaria.nyfspetshop.event;
+
+import com.nyfaria.nyfspetshop.init.EntityInit;
+import net.minecraftforge.event.entity.EntityAttributeCreationEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+
+public class CommonModEvents {
+
+
+    @SubscribeEvent
+    public static void attribs(EntityAttributeCreationEvent e) {
+        EntityInit.attributeSuppliers.forEach(p -> e.put(p.entityTypeSupplier().get(), p.factory().get().build()));
+    }
+}