+++ Common/changelog.md
@@ -1,5 +1,7 @@
-### Version 1.0.1
-- fixed a particular bug
-
 ### Version 1.0.0
 - Initial release.
+- Added Pet Shop Villager
+- Added Pet Shop to Plains Village
+- Added 2 Dog Pets
+- Added 2 Bird Pets
+- Added 2 Cat Pets